refactor(workout-plan): extract swap/replace cluster to workout-plan-replacement.js (WP3.4e)
Move-only split of the swap/replace-exercise cluster out of the workout-plan
monolith into a new workout-plan-replacement.js module. Moved verbatim:
getSwapButtonLoadingMarkup (private) and handleSwapExercise (exported).

Plain one-directional import (WP3.4c-style), NOT callback injection:
handleSwapExercise updates the row in place and never calls fetchWorkoutPlan(),
so the new module imports its deps directly — updateRowMetadata/updateCachedExercise
from workout-plan-table.js, notifyVolumeAffectingPlanChange from
workout-plan-events.js, and the pure helpers buildReplacePayload/resolveSwapErrorToast
from workout-plan-helpers.js — with no back-import into the monolith (no circular
import). The monolith imports handleSwapExercise and keeps injecting it into the
table via configureWorkoutPlanTable({ onSwap: handleSwapExercise }); the now-unused
buildReplacePayload/resolveSwapErrorToast and updateRowMetadata/updateCachedExercise
imports are pruned from the monolith.

No behavior/API/DB/calculation/response-shape/event-timing change. Moved bodies
byte-identical to HEAD (only 'export' prepended to the handleSwapExercise def);
monolith remainder proven == HEAD minus moved/pruned lines plus one new import.

Gates: Vitest 88; E2E replace-exercise-errors + workout-plan + exercise-interactions
+ superset-edge-cases + fatigue-context 77 passed; learned-calibration 8/8 isolated.
